Fix Shape parameter real-time modulation
The Shape parameter was not responding to real-time changes during playback
because it was still being treated as an integer (0 or 1) instead of the
new continuous float (0.0-1.0) value.

Changes:
- Updated Synth::setOscillatorState signature: int shape → float shape
- Removed old discrete shape mapping logic (0→0.5, 1→0.9)
- Shape now passes through directly from UI to oscillator
- Updated main.cpp to read shape as float instead of int

Now the full 5-shape continuum responds immediately to parameter changes
during playback (Sine → Triangle → Saw → Square → Pulse).

+++ wakefield/src/main.cpp
@@ -518,7 +518,7 @@ int audioCallback(void* outputBuffer, void* /*inputBuffer*/,
         // Update per-oscillator parameters (oscillator 1 uses smoothed values)
         for (int oscIndex = 0; oscIndex < OSCILLATORS_PER_VOICE; ++oscIndex) {
             BrainwaveMode mode = static_cast<BrainwaveMode>(synthParams->getOscMode(oscIndex));
-            int shape = synthParams->getOscShape(oscIndex);
+            float shape = synthParams->getOscShape(oscIndex);
             float baseFreq = (oscIndex == 0) ? smoothedOscillatorFreq : synthParams->getOscFrequency(oscIndex);
             float morph = (oscIndex == 0) ? smoothedOscillatorMorph : synthParams->getOscMorph(oscIndex);
             float duty = (oscIndex == 0) ? smoothedOscillatorDuty : synthParams->getOscDuty(oscIndex);
